Giving permission logged in user to create edit delete ingridient

+++ app/controllers/ingridients_controller.rb
@@ -2,45 +2,61 @@ class IngridientsController < ApplicationController
   before_action :find_ingridient, only: [:edit, :update, :destroy]
 
   def new
-    @recipe = Recipe.find(params[:recipe_id])
-    @ingridient = @recipe.ingridients.build
+    if logged_in? 
+      @recipe = Recipe.find(params[:recipe_id])
+      @ingridient = @recipe.ingridients.build
+    else
+      flash[:danger] = 'You have to login!'
+      redirect_to root_path
+    end
   end
 
   def create
-    recipe = Recipe.find(params[:recipe_id])
-    ingridient = recipe.ingridients.create(ingridient_params)
+    if logged_in? 
+      recipe = Recipe.find(params[:recipe_id])
+      ingridient = recipe.ingridients.create(ingridient_params)
 
-    if ingridient.save
-      redirect_to recipe
+      if ingridient.save
+        redirect_to recipe
+      else
+        render :new
+      end
     else
-      render :new
+      flash[:danger] = 'You have to login!'
+      redirect_to root_path
     end
   end
 
   def edit
-    @ingridient = Ingridient.find(params[:id])
-    @recipe = @ingridient.recipe
-  end  
+    unless logged_in?
+      @ingridient = Ingridient.find(params[:id])
+      @recipe = @ingridient.recipe
+    end
+  end
 
   def update
-    @ingridient = Ingridient.find(params[:id])
-    @recipe = @ingridient.recipe
-
-    if @ingridient.update(ingridient_params)
-      redirect_to @recipe
+    if logged_in?
+      if @ingridient.update(ingridient_params)
+        redirect_to @recipe
+      else
+        render :edit
+      end
     else
-      render :edit
+      flash[:danger] = 'You have to login!'
+      redirect_to root_path
     end
   end
 
   def destroy
-    @ingridient = Ingridient.find(params[:id])
-    @recipe = @ingridient.recipe
-
-    if @ingridient.destroy
-      redirect_to @recipe
+    if logged_in?
+      if @ingridient.destroy
+        redirect_to @recipe
+      else
+        render @recipe
+      end
     else
-      render @recipe
+      flash[:danger] = 'You have to login!'
+      redirect_to root_path
     end
   end
